✨ feat: lucky wheel and color
